create endpoints for transportation

+++ backend/src/main/java/com/chinguv45tier3team53/tripplannerrest/dao/TransportationRepository.java
@@ -3,5 +3,5 @@ package com.chinguv45tier3team53.tripplannerrest.dao;
 import com.chinguv45tier3team53.tripplannerrest.entity.Transportation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransportationRepository extends JpaRepository<Transportation, Integer> {
+public interface TransportationRepository extends JpaRepository<Transportation, Long> {
 }
